fix KLE data, was incomplete before
In the previous commit, I missed a subtle bug. The XML file containing
data (not structure) did not contain all the elements from KLE
Emneplan. The reason is, it was exported from MW after importing as
CSV. The MW XML export appears to be limited to a number of pages much
lower than the total amount of KLE Emneplan elements, and it does not
warn you when adding the category to the export list.

This commit rectifies that by abandoning the XML import for data in
favor of importing the actual generated CSV files. Of course this
requires ExternalData be installed, but that's a prerequisite anyhow.