preferve Free text in wiki pages (#62)
This commmit is a candidate to the ticket #62 about not
overwriting/removing Free Text in wiki pages when pulling Geodata tables
from Geocloud2.

A new function get_freetext_from_guid is introduced; it's fairly
expensive, but that's how it is for now with MediaWiki. Importantly, be
aware that we've hardcoded the method in which we isolate the Free Text
contents, by stripping out the Template:Geodata call. There are many
ways in which this can break. Most notably, if we add or remove fields
from the template without updating this script, it will break.

Another quite important consideration: we're losing revision history
with this method. It isn't the biggest job in the world to modify the
script to preserve page history (remove the call to delete the category
and mmanually prune deleted tables) but for now, we're discarding it in
the name of results.